Extract CSS reference helper
