+++ src/CodeIndex/Indexer/ReferenceExtractor.cs
@@ -31,14 +31,6 @@ public static class ReferenceExtractor
         "zig", "css"
     ];
 
-    private static readonly Regex ScssVariableReferenceRegex = new(
-        @"(?<![\w$])\$(?<name>[A-Za-z_][\w-]*)",
-        RegexOptions.Compiled);
-
-    private static readonly Regex ScssExtendReferenceRegex = new(
-        @"@extend\s+(?<name>[%.][A-Za-z_][\w-]*)",
-        RegexOptions.Compiled);
-
     // Batch jump targets can appear as direct commands, chained commands, or inline `if`
     // forms, including comparison-based conditions such as `if /i "%a%"=="b" goto :X`.
     // batch のジャンプ先は、直書き・連結コマンド・`if` 併用の inline 形として現れうる。
@@ -276,18 +268,6 @@ public static class ReferenceExtractor
     private static readonly Regex PhpObjectMemberAccessRegex = new(
         @"(?:\?->|->)\s*(?<name>[A-Za-z_]\w*)(?!\s*\()",
         RegexOptions.Compiled | RegexOptions.CultureInvariant);
-    private static readonly Regex CssCustomPropertyReferenceRegex = new(@"\bvar\(\s*--(?<name>[\w-]+)", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex CssAnimationNameReferenceRegex = new(@"\banimation-name\s*:\s*(?<name>[\w-]+)", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex CssAnimationShorthandValueRegex = new(@"\banimation\s*:\s*(?<value>[^;{}]+)", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex CssClassSelectorReferenceRegex = new(@"(?<![A-Za-z0-9_-])\.(?<name>[\w-]+)", RegexOptions.Compiled);
-    private static readonly HashSet<string> CssAnimationShorthandIgnoredTokens = new(StringComparer.OrdinalIgnoreCase)
-    {
-        "ease", "ease-in", "ease-out", "ease-in-out", "linear",
-        "step-start", "step-end", "cubic-bezier", "steps",
-        "infinite", "normal", "reverse", "alternate", "alternate-reverse",
-        "none", "forwards", "backwards", "both", "running", "paused",
-        "initial", "inherit", "unset", "revert", "revert-layer",
-    };
     // SQL-specific single-quoted string stripper: preserve identifier quoting (`[...]`, `` `...` ``,
     // and ANSI `"..."`) so the SQL graph path can still see real object names while literal payloads
     // stay masked.
@@ -1637,7 +1617,7 @@ public static class ReferenceExtractor
             }
             else if (language == "css")
             {
-                EmitCssReferences(
+                CssReferenceExtractor.EmitCss(
                     preparedLine,
                     context,
                     lineNumber,
@@ -2043,7 +2023,7 @@ public static class ReferenceExtractor
 
             if (language == "css")
             {
-                EmitCssScssReferences(
+                CssReferenceExtractor.EmitScss(
                     preparedLine,
                     references,
                     seen,
@@ -2171,7 +2151,7 @@ public static class ReferenceExtractor
 
             if (language == "css")
             {
-                EmitCssScssReferences(
+                CssReferenceExtractor.EmitScss(
                     preparedLine,
                     references,
                     seen,
@@ -2863,59 +2843,6 @@ public static class ReferenceExtractor
         });
     }
 
-    private static void EmitCssReferences(
-        string preparedLine,
-        string context,
-        int lineNumber,
-        List<ReferenceRecord> references,
-        HashSet<string> seen,
-        long fileId,
-        HashSet<string>? definitionNames,
-        SymbolRecord? container)
-    {
-        foreach (Match match in CssCustomPropertyReferenceRegex.Matches(preparedLine))
-        {
-            var nameGroup = match.Groups["name"];
-            if (definitionNames != null && definitionNames.Contains(nameGroup.Value))
-                continue;
-
-            AddReference(references, seen, fileId, nameGroup.Value, nameGroup.Index, "reference", context, lineNumber, container);
-        }
-
-        foreach (Match match in CssAnimationNameReferenceRegex.Matches(preparedLine))
-        {
-            var nameGroup = match.Groups["name"];
-            if (definitionNames != null && definitionNames.Contains(nameGroup.Value))
-                continue;
-
-            AddReference(references, seen, fileId, nameGroup.Value, nameGroup.Index, "reference", context, lineNumber, container);
-        }
-
-        foreach (Match match in CssAnimationShorthandValueRegex.Matches(preparedLine))
-        {
-            EmitCssAnimationShorthandReferences(
-                match.Groups["value"].Value,
-                match.Groups["value"].Index,
-                context,
-                lineNumber,
-                references,
-                seen,
-                fileId,
-                definitionNames,
-                container);
-        }
-
-        EmitCssClassSelectorReferences(
-            preparedLine,
-            context,
-            lineNumber,
-            references,
-            seen,
-            fileId,
-            definitionNames,
-            container);
-    }
-
     private static bool IsJsxFilePath(string? path)
     {
         if (string.IsNullOrWhiteSpace(path))
@@ -2924,327 +2851,6 @@ public static class ReferenceExtractor
         var extension = Path.GetExtension(path);
         return string.Equals(extension, ".jsx", StringComparison.OrdinalIgnoreCase)
             || string.Equals(extension, ".tsx", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static void EmitCssAnimationShorthandReferences(
-        string value,
-        int valueIndex,
-        string context,
-        int lineNumber,
-        List<ReferenceRecord> references,
-        HashSet<string> seen,
-        long fileId,
-        HashSet<string>? definitionNames,
-        SymbolRecord? container)
-    {
-        var segmentStart = 0;
-        var parenDepth = 0;
-        for (var i = 0; i <= value.Length; i++)
-        {
-            if (i < value.Length)
-            {
-                var ch = value[i];
-                if (ch == '(')
-                {
-                    parenDepth++;
-                    continue;
-                }
-
-                if (ch == ')' && parenDepth > 0)
-                {
-                    parenDepth--;
-                    continue;
-                }
-
-                if (ch != ',' || parenDepth > 0)
-                    continue;
-            }
-
-            EmitCssAnimationShorthandSegmentReference(
-                value,
-                valueIndex,
-                segmentStart,
-                i,
-                context,
-                lineNumber,
-                references,
-                seen,
-                fileId,
-                definitionNames,
-                container);
-            segmentStart = i + 1;
-        }
-    }
-
-    private static void EmitCssAnimationShorthandSegmentReference(
-        string value,
-        int valueIndex,
-        int segmentStart,
-        int segmentEnd,
-        string context,
-        int lineNumber,
-        List<ReferenceRecord> references,
-        HashSet<string> seen,
-        long fileId,
-        HashSet<string>? definitionNames,
-        SymbolRecord? container)
-    {
-        var cursor = segmentStart;
-        while (cursor < segmentEnd && char.IsWhiteSpace(value[cursor]))
-            cursor++;
-
-        while (cursor < segmentEnd)
-        {
-            var tokenStart = cursor;
-            while (cursor < segmentEnd && !char.IsWhiteSpace(value[cursor]))
-                cursor++;
-
-            var token = value[tokenStart..cursor];
-            if (!IsCssAnimationNameToken(token))
-                continue;
-            if (definitionNames != null && definitionNames.Contains(token))
-                return;
-
-            AddReference(references, seen, fileId, token, valueIndex + tokenStart, "reference", context, lineNumber, container);
-            return;
-        }
-    }
-
-    private static void EmitCssClassSelectorReferences(
-        string preparedLine,
-        string context,
-        int lineNumber,
-        List<ReferenceRecord> references,
-        HashSet<string> seen,
-        long fileId,
-        HashSet<string>? definitionNames,
-        SymbolRecord? container)
-    {
-        var segmentStart = 0;
-        while (segmentStart < preparedLine.Length)
-        {
-            var braceIndex = preparedLine.IndexOf('{', segmentStart);
-            var segmentEnd = braceIndex >= 0 ? braceIndex : preparedLine.Length;
-            var trimmedStart = segmentStart;
-            while (trimmedStart < segmentEnd && char.IsWhiteSpace(preparedLine[trimmedStart]))
-                trimmedStart++;
-
-            if (trimmedStart < segmentEnd && preparedLine[trimmedStart] != '@')
-            {
-                var selectorSegment = preparedLine[trimmedStart..segmentEnd];
-                foreach (var (partStart, partEnd) in EnumerateCssSelectorListSegments(selectorSegment))
-                {
-                    var selectorPart = selectorSegment[partStart..partEnd];
-                    if (!ContainsCssClassSelectorReferenceCandidate(selectorPart))
-                        continue;
-
-                    var selectorPartTrimStart = 0;
-                    while (selectorPartTrimStart < selectorPart.Length && char.IsWhiteSpace(selectorPart[selectorPartTrimStart]))
-                        selectorPartTrimStart++;
-
-                    var selectorPartBody = selectorPart[selectorPartTrimStart..];
-                    foreach (Match match in CssClassSelectorReferenceRegex.Matches(selectorPartBody))
-                    {
-                        var nameGroup = match.Groups["name"];
-                        var name = "." + nameGroup.Value;
-                        if (definitionNames != null && definitionNames.Contains(name))
-                            continue;
-
-                        AddReference(
-                            references,
-                            seen,
-                            fileId,
-                            name,
-                            trimmedStart + partStart + selectorPartTrimStart + match.Groups["name"].Index - 1,
-                            "reference",
-                            context,
-                            lineNumber,
-                            container);
-                    }
-                }
-            }
-
-            if (braceIndex < 0)
-                break;
-
-            segmentStart = braceIndex + 1;
-        }
-    }
-
-    private static IEnumerable<(int Start, int End)> EnumerateCssSelectorListSegments(string selectorSegment)
-    {
-        var segmentStart = 0;
-        var parenDepth = 0;
-        var bracketDepth = 0;
-
-        for (var index = 0; index < selectorSegment.Length; index++)
-        {
-            var ch = selectorSegment[index];
-            if (ch == '(')
-            {
-                parenDepth++;
-                continue;
-            }
-
-            if (ch == ')' && parenDepth > 0)
-            {
-                parenDepth--;
-                continue;
-            }
-
-            if (ch == '[')
-            {
-                bracketDepth++;
-                continue;
-            }
-
-            if (ch == ']' && bracketDepth > 0)
-            {
-                bracketDepth--;
-                continue;
-            }
-
-            if (ch == ',' && parenDepth == 0 && bracketDepth == 0)
-            {
-                yield return (segmentStart, index);
-                segmentStart = index + 1;
-            }
-        }
-
-        yield return (segmentStart, selectorSegment.Length);
-    }
-
-    private static bool ContainsCssClassSelectorReferenceCandidate(string selectorPart)
-    {
-        var bracketDepth = 0;
-        char quote = '\0';
-        for (var index = 0; index < selectorPart.Length; index++)
-        {
-            var ch = selectorPart[index];
-            if (quote != '\0')
-            {
-                if (ch == quote && (index == 0 || selectorPart[index - 1] != '\\'))
-                    quote = '\0';
-                continue;
-            }
-
-            if (ch is '\'' or '"')
-            {
-                quote = ch;
-                continue;
-            }
-
-            if (ch == '[')
-            {
-                bracketDepth++;
-                continue;
-            }
-
-            if (ch == ']' && bracketDepth > 0)
-            {
-                bracketDepth--;
-                continue;
-            }
-
-            if (bracketDepth == 0 && ch == '.')
-                return true;
-        }
-
-        return false;
-    }
-
-    private static bool IsCssAnimationNameToken(string token)
-    {
-        if (string.IsNullOrWhiteSpace(token))
-            return false;
-
-        if (CssAnimationShorthandIgnoredTokens.Contains(token))
-            return false;
-        if (token.IndexOf('(') >= 0 || token.IndexOf(')') >= 0 || token.IndexOf(',') >= 0
-            || token.IndexOf('/') >= 0 || token.IndexOf(':') >= 0 || token.IndexOf(';') >= 0)
-            return false;
-        if (IsCssAnimationTimeToken(token) || IsCssAnimationNumberToken(token))
-            return false;
-        if (token.StartsWith("--", StringComparison.Ordinal))
-            return false;
-        if (!(char.IsLetter(token[0]) || token[0] == '_' || token[0] == '-'))
-            return false;
-        if (token[0] == '-' && token.Length > 1 && (token[1] == '-' || char.IsDigit(token[1])))
-            return false;
-
-        for (var i = 1; i < token.Length; i++)
-        {
-            if (char.IsLetterOrDigit(token[i]) || token[i] == '_' || token[i] == '-')
-                continue;
-            return false;
-        }
-
-        return true;
-    }
-
-    private static bool IsCssAnimationTimeToken(string token)
-    {
-        if (token.Length < 2)
-            return false;
-
-        var unitLength = token.EndsWith("ms", StringComparison.OrdinalIgnoreCase)
-            ? 2
-            : token.EndsWith("s", StringComparison.OrdinalIgnoreCase)
-                ? 1
-                : 0;
-        if (unitLength == 0 || token.Length == unitLength)
-            return false;
-
-        var numberPart = token[..^unitLength];
-        var sawDigit = false;
-        var sawDot = false;
-        foreach (var ch in numberPart)
-        {
-            if (char.IsDigit(ch))
-            {
-                sawDigit = true;
-                continue;
-            }
-
-            if (ch == '.' && !sawDot)
-            {
-                sawDot = true;
-                continue;
-            }
-
-            return false;
-        }
-
-        return sawDigit;
-    }
-
-    private static bool IsCssAnimationNumberToken(string token)
-    {
-        if (token.Length == 0 || token.IndexOfAny(['(', ')', ',', '/', ':', ';']) >= 0)
-            return false;
-        if (!(char.IsDigit(token[0]) || token[0] == '.'))
-            return false;
-
-        var sawDigit = false;
-        var sawDot = false;
-        foreach (var ch in token)
-        {
-            if (char.IsDigit(ch))
-            {
-                sawDigit = true;
-                continue;
-            }
-
-            if (ch == '.' && !sawDot)
-            {
-                sawDot = true;
-                continue;
-            }
-
-            return false;
-        }
-
-        return sawDigit;
     }
 
     private static void NormalizeSqlIdentifier(
@@ -8487,49 +8093,6 @@ public static class ReferenceExtractor
         return line;
     }
 
-    private static void EmitCssScssReferences(
-        string preparedLine,
-        List<ReferenceRecord> references,
-        HashSet<string> seen,
-        long fileId,
-        string context,
-        int lineNumber,
-        SymbolRecord? container)
-    {
-        foreach (Match match in ScssVariableReferenceRegex.Matches(preparedLine))
-        {
-            var nameGroup = match.Groups["name"];
-            if (ShouldSkipScssVariableReference(preparedLine, nameGroup.Index))
-                continue;
-
-            AddReference(
-                references,
-                seen,
-                fileId,
-                nameGroup.Value,
-                nameGroup.Index,
-                "call",
-                context,
-                lineNumber,
-                container);
-        }
-
-        foreach (Match match in ScssExtendReferenceRegex.Matches(preparedLine))
-        {
-            var nameGroup = match.Groups["name"];
-            AddReference(
-                references,
-                seen,
-                fileId,
-                nameGroup.Value,
-                nameGroup.Index,
-                "call",
-                context,
-                lineNumber,
-                container);
-        }
-    }
-
     private static void EmitPhpStaticAccessReferences(
         string preparedLine,
         List<ReferenceRecord> references,
@@ -8610,29 +8173,6 @@ public static class ReferenceExtractor
                 lineNumber,
                 container);
         }
-    }
-
-    private static bool ShouldSkipScssVariableReference(string preparedLine, int variableIndex)
-    {
-        var trimmed = preparedLine.TrimStart();
-        if (trimmed.StartsWith("$", StringComparison.Ordinal))
-        {
-            var declarationColonIndex = preparedLine.IndexOf(':', variableIndex);
-            if (declarationColonIndex >= 0)
-                return true;
-        }
-
-        if (trimmed.StartsWith("@mixin", StringComparison.Ordinal)
-            || trimmed.StartsWith("@function", StringComparison.Ordinal))
-        {
-            var braceIndex = preparedLine.IndexOf('{');
-            if (braceIndex < 0)
-                return true;
-            if (variableIndex < braceIndex)
-                return true;
-        }
-
-        return false;
     }
 
     private static string NormalizeCSharpQualifiedSegments(
